backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Computer Vision
opencv-python-headless==4.9.0.80
easyocr==1.7.1
ultralytics==8.1.0
Pillow==10.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Async
aiofiles==23.2.1
